Updates device saving and node device handling
Refines the device saving mechanism by providing a default value for the prefix, enhancing flexibility.

Ensures correct type conversion when retrieving the size of node device data, preventing potential data loss.

Improves node device pointer handling by using `auto` for type inference, leading to more maintainable code.

+++ PAC/common/device/node_dev.cpp
@@ -40,7 +40,7 @@ int node_dev::save_device_ex( char* buff )
     auto res_n = fmt::format_to_n( buff, MAX_COPY_SIZE,
         "WEB = {}, STARTUP = {}, IP = '{}', ",
         web_value, startup_value, get_ip() );
-    return res_n.size;
+    return static_cast<int>( res_n.size );
     }
 //-----------------------------------------------------------------------------
 const char* node_dev::get_name_in_Lua() const
